Moving RGBDepth to shared_data_structs
