+++ droidlet/perception/robot/__init__.py
@@ -13,10 +13,10 @@ from droidlet.perception.robot.handlers import (
     ObjectDeduplicationHandler,
     Detection,
     WorldObject,
-    RGBDepth,
     Human,
     HumanKeypoints,
 )
+from droidlet.shared_data_struct.shared_data_structs import RGBDepth
 from droidlet.perception.robot.perception import Perception, SlowPerception
 from droidlet.perception.robot.self_perception import SelfPerception
