+++ droidlet/perception/robot/handlers/detector.py
@@ -17,7 +17,8 @@ from detectron2.engine.defaults import DefaultPredictor
 
 import droidlet.memory.robot.loco_memory as loco_memory
 
-from .core import AbstractHandler, WorldObject, RGBDepth
+from .core import AbstractHandler, WorldObject
+from droidlet.shared_data_struct.shared_data_structs import RGBDepth
 from ..detectron.detector.utils import get_predictor
 from ..detectron.detector.visualizer import LocobotVisualizer
 from droidlet.perception.robot.perception_helpers import get_color_tag
